+++ src/main/java/com/lms/voting/service/VotingService.java
@@ -21,6 +21,6 @@ public interface VotingService {
 
     PartyVoteSummary getTotalVotesByParty(Integer partyId);
 
-    Integer getTotalCountVoter();
+    Long getTotalCountVoter();
 
 }
